# daemon/docs/swagger.yaml
basePath: /
definitions:
...
  internal_handler_wallpaperhandler.WallpaperCurrentResponse:
    properties:
      backend:
        type: string
      image_id:
        type: integer
      image_name:
        type: string
      image_path:
        type: string
      mode:
        type: string
      monitors:
        items:
          $ref: '#/definitions/internal_handler_wallpaperhandler.WallpaperCurrentSlot'
        type: array
      set_at:
        type: string
    type: object
  internal_handler_wallpaperhandler.WallpaperCurrentSlot:
    properties:
      image_id:
        type: integer
      image_name:
        type: string
      image_path:
        type: string
      monitor_name:
        type: string
      set_at:
        type: string
    type: object
  internal_handler_wallpaperhandler.setWallpaperRequest:
    properties:
      image_id:
        type: integer
      mode:
        $ref: '#/definitions/waypaper-engine_daemon_internal_monitor.MonitorMode'
      monitor:
        type: string
      monitors:
        items:
          type: string
        type: array
    type: object
...
  waypaper-engine_daemon_internal_handler_httpjson.APIError:
    properties:
      code:
        type: integer
      details:
        type: string
      error:
        type: string
      error_code:
        type: string
      meta:
        additionalProperties: {}
        type: object
    type: object
...
  waypaper-engine_daemon_internal_monitor.MonitorMode:
    enum:
    - individual
    - clone
    - extend
    type: string
    x-enum-varnames:
    - ModeIndividual
    - ModeClone
    - ModeExtend
...
info:
  contact: {}
  description: HTTP API served over a Unix socket for the waypaper-engine wallpaper
    manager.
  title: waypaper-engine daemon
  version: 3.0.0
paths:
...
  /images/history:
    delete:
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
      summary: Clear wallpaper history
      tags:
      - images
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
      summary: Get wallpaper history
      tags:
      - images
...
  /wallpaper/current:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler_wallpaperhandler.WallpaperCurrentResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
      summary: Get current wallpaper state
      tags:
      - wallpaper
  /wallpaper/random:
    post:
      parameters:
      - description: Optional monitor/mode
        in: body
        name: body
        schema:
          additionalProperties: true
          type: object
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
      summary: Set random wallpaper
      tags:
      - wallpaper
  /wallpaper/set:
    post:
      parameters:
      - description: Set request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler_wallpaperhandler.setWallpaperRequest'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
      summary: Set wallpaper
      tags:
      - wallpaper
swagger: "2.0"